backend/requirements.txt
fastapi
uvicorn
pydantic
firebase-admin
cloudinary
torch
torchvision
torchaudio
scikit-learn
pandas
numpy
sentence-transformers
faiss-cpu
python-dotenv
requests
pytest
